Add actual deduping for all videos from any source, Move media parsing logic to ModuleDetails
